Rework what happens if pool is empty to be more stable
Before, when the pool was empty (which often happens during resizing)
the configuration was put on hold until we got a buffer back. This
created crashes that probably came from trying to act on wrongly sized
buffers. There was also stutters that probably came due to not
responding to the configure events fast enough